separate internal and external public processors
