+++ oraclecloud-serde-internal-processor/src/main/resources/META-INF/services/io.micronaut.inject.visitor.TypeElementVisitor
@@ -1,0 +1,1 @@
+io.micronaut.oraclecloud.httpclient.netty.visitor.OciSdkModelSerdeVisitor
