+++ oraclecloud-serde-internal-processor/src/test/groovy/io/micronaut/oraclecloud/httpclient/netty/visitor/OCISdkModelSerdeVisitorSpec.groovy
