+++ oraclecloud-serde-processor/src/main/java/io/micronaut/oraclecloud/sdk/processor/SdkImportVisitor.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micronaut.oraclecloud.httpclient.netty.visitor;
+package io.micronaut.oraclecloud.sdk.processor;
 
 import io.micronaut.context.annotation.Bean;
 import io.micronaut.context.annotation.Factory;
