+++ settings.gradle.kts
@@ -30,6 +30,7 @@ include("oraclecloud-function-http-test")
 include("oraclecloud-httpclient-apache-http-core")
 include("oraclecloud-httpclient-netty")
 include("oraclecloud-serde-processor")
+include("oraclecloud-serde-internal-processor")
 include("oraclecloud-logging")
 include("oraclecloud-micrometer")
 include("oraclecloud-oke-kubernetes-client")
